Working axpb on GPU too.

+++ src/axpb.cu
@@ -10,20 +10,20 @@
 #include <stdlib.h>
 #include <math.h>
 
-__global__ void Vec_add(float x[], float y[], float z[], int n) {
-    int thread_id = threadIdx.x;
-    if (thread_id < n){
-        z[thread_id] = x[thread_id] + y[thread_id];
-    }
+__global__ void Vec_add(double x[], double y[], double z[], int n) {
+  int thread_id = threadIdx.x;
+  if (thread_id < n){
+    z[thread_id] = x[thread_id] + y[thread_id];
+  }
 }
 
 
-extern "C" void axpb_cpp_cuda(float h_x[], float h_y[], float h_z[], int n) {
-  float *d_x, *d_y, *d_z;
+extern "C" void axpb_cpp_cuda(double h_x[], double h_y[], double h_z[], int n) {
+  double *d_x, *d_y, *d_z;
   size_t size;
 
   /* Define vector length */
-  size = n*sizeof(float);
+  size = n*sizeof(double);
 
   // Print original vectors.
   printf("h_x = ");
